ショップアイテム購入処理の追加

+++ program/game/Actor/Actor.h
@@ -120,7 +120,7 @@ protected:
 	int hpVar_gh = 0;
 
 	const float VARWIDTH = 30;
-	double nowHpVarWidth = nowHp / hp;
+	double nowHpVarWidth = static_cast<double>(nowHp / hp);
 
 
 	int actSpeed = 20;
